requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
qdrant-client==1.9.1
openai==1.6.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
pytest==7.4.0
httpx==0.25.0